models renamed to service. utils path changed.
