+++ service/contacts.js
@@ -1,6 +1,6 @@
 const fs = require('fs/promises')
 const path = require("path");
-const createId = require('./utils/idGenerator');
+const createId = require('../utils/idGenerator');
 
 const contactsPath = path.join(__dirname, "contacts.json");
